+++ protomol/src/protomol/output/OutputXYZTrajectoryPos.h
@@ -24,8 +24,8 @@ namespace ProtoMol {
   private:
     Output *doMake(const std::vector<Value> &values) const;
     void doInitialize();
-    void doRun(int step);
-    void doFinalize(int step);
+    void doRun(long step);
+    void doFinalize(long step);
 
     //  From class Makeabl
   public:
